chore(metrics): clean up MetricsInterface & API links
