+++ packages/metrics/src/middleware/middy.ts
@@ -7,31 +7,40 @@ import type { Metrics } from '../Metrics.js';
 import type { ExtraOptions } from '../types/Metrics.js';
 
 /**
- * A middy middleware automating capture of metadata and annotations on segments or subsegments for a Lambda Handler.
+ * A Middy.js middleware automating capture of Amazon CloudWatch metrics.
  *
- * Using this middleware on your handler function will automatically flush metrics after the function returns or throws an error.
- * Additionally, you can configure the middleware to easily:
- * * ensure that at least one metric is emitted before you flush them
- * * capture a `ColdStart` a metric
- * * set default dimensions for all your metrics
+ * This middleware is compatible with `@middy/core@3.x` and above.
+ *
+ * The middleware automatically flushes metrics after the handler function returns or throws an error,
+ * so you don't need to call {@link Metrics.publishStoredMetrics | `publishStoredMetrics()`} manually.
  *
  * @example
  * ```typescript
- * import { Metrics } from '@aws-lambda-powertools/metrics';
+ * import { Metrics, MetricUnit } from '@aws-lambda-powertools/metrics';
  * import { logMetrics } from '@aws-lambda-powertools/metrics/middleware';
  * import middy from '@middy/core';
  *
- * const metrics = new Metrics({ namespace: 'serverlessAirline', serviceName: 'orders' });
+ * const metrics = new Metrics({
+ *   namespace: 'serverlessAirline',
+ *   serviceName: 'orders'
+ * });
  *
- * const lambdaHandler = async (_event: any, _context: any) => {
- *   ...
- * };
- *
- * export const handler = middy(lambdaHandler).use(logMetrics(metrics));
+ * export const handler = middy(async () => {
+ *   metrics.addMetadata('request_id', event.requestId);
+ *   metrics.addMetric('successfulBooking', MetricUnit.Count, 1);
+ * }).use(logMetrics(metrics, {
+ *   captureColdStartMetric: true,
+ *   throwOnEmptyMetrics: true,
+ * }));
  * ```
  *
+ * You can configure the middleware with the following options:
+ * - `captureColdStartMetric`: Whether to capture a `ColdStart` metric
+ * - `defaultDimensions`: Default dimensions to add to all metrics
+ * - `throwOnEmptyMetrics`: Whether to throw an error if no metrics are emitted
+ *
  * @param target - The Metrics instance to use for emitting metrics
- * @param options - (_optional_) Options for the middleware
+ * @param options - Options to configure the middleware, see {@link ExtraOptions}
  */
 const logMetrics = (
   target: Metrics | Metrics[],
@@ -57,7 +66,7 @@ const logMetrics = (
       const { throwOnEmptyMetrics, defaultDimensions, captureColdStartMetric } =
         options;
       if (throwOnEmptyMetrics) {
-        metrics.throwOnEmptyMetrics();
+        metrics.setThrowOnEmptyMetrics(throwOnEmptyMetrics);
       }
       if (defaultDimensions !== undefined) {
         metrics.setDefaultDimensions(defaultDimensions);
